+++ recipes/upgrade.rb
@@ -22,11 +22,6 @@
 # https://confluence.atlassian.com/display/STASH/Stash+upgrade+guide
 #
 
-stash_data_bag = Chef::EncryptedDataBagItem.load("stash","stash")
-stash_configuration_info = stash_data_bag[node.chef_environment]['configuration']
-stash_database_info = stash_data_bag[node.chef_environment]['database']
-stash_tomcat_info = stash_data_bag[node.chef_environment]['tomcat']
-
 service "stash" do
   action :stop
 end
@@ -41,71 +36,12 @@ execute "Backing up Stash Install Directory" do
   only_if { node['stash']['backup_install'] }
 end
 
-remote_file "#{Chef::Config[:file_cache_path]}/atlassian-stash-#{node['stash']['version']}.tar.gz" do
-  source    node['stash']['url']
-  checksum  node['stash']['checksum']
-  mode      "0644"
-  action    :create_if_missing
+directory node['stash']['install_path'] do
+  recursive true
+  action :delete
 end
 
-execute "Extracting Stash #{node['stash']['version']} for Upgrade" do
-  cwd Chef::Config[:file_cache_path]
-  command <<-COMMAND
-    tar -zxf atlassian-stash-#{node['stash']['version']}.tar.gz
-    chown -R #{node['stash']['run_user']} atlassian-stash-#{node['stash']['version']}
-    rm -rf #{node['stash']['install_path']}
-    mv atlassian-stash-#{node['stash']['version']} #{node['stash']['install_path']}
-  COMMAND
-end
-
-if stash_database_info['type'] == "mysql"
-  remote_file "#{Chef::Config[:file_cache_path]}/mysql-connector-java-#{node['stash']['mysql']['connector']['version']}.tar.gz" do
-    source    node['stash']['mysql']['connector']['url']
-    checksum  node['stash']['mysql']['connector']['checksum']
-    mode      "0644"
-    action    :create_if_missing
-  end
-
-  execute "Extracting MySQL Connector Java #{node['stash']['mysql']['connector']['version']}" do
-    cwd Chef::Config[:file_cache_path]
-    command <<-COMMAND
-      tar -zxf mysql-connector-java-#{node['stash']['mysql']['connector']['version']}.tar.gz
-      chown -R #{node['stash']['run_user']} mysql-connector-java-#{node['stash']['mysql']['connector']['version']}
-      mv mysql-connector-java-#{node['stash']['mysql']['connector']['version']}/mysql-connector-java-#{node['stash']['mysql']['connector']['version']}-bin.jar #{node['stash']['install_path']}/lib
-    COMMAND
-    creates "#{node['stash']['install_path']}/lib/mysql-connector-java-#{node['stash']['mysql']['connector']['version']}-bin.jar"
-  end
-end
-
-template "#{node['stash']['install_path']}/bin/setenv.sh" do
-  source "setenv.sh.erb"
-  owner  node['stash']['run_user']
-  mode   "0755"
-end
-
-template "#{node['stash']['install_path']}/conf/server.xml" do
-  source "server.xml.erb"
-  owner  node['stash']['run_user']
-  mode   "0640"
-  variables :tomcat => stash_tomcat_info
-end
-
-template "#{node['stash']['install_path']}/conf/web.xml" do
-  source "web.xml.erb"
-  owner  node['stash']['run_user']
-  mode   "0644"
-end
-
-template "#{node['stash']['install_path']}/stash-config.properties" do
-  source "stash-config.properties.erb"
-  owner  node['stash']['run_user']
-  mode   "0644"
-  variables :database => stash_database_info
-end
-
-service "stash" do
-  action :start
-end
+include_recipe "stash"
 
 ruby_block "remove_recipe_stash_upgrade" do
   block { node.run_list.remove("recipe[stash::upgrade]") }